requirements.txt
asgiref==3.2.10
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Cython==0.29.21
Django==3.1.2
fpdf==1.7.2
google-api-core==1.22.4
google-auth==1.22.1
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
idna==2.10
install==1.3.4
libcst==0.3.13
mypy-extensions==0.4.3
numpy==1.19.2
opencv-python==4.4.0.44
pdf2image==1.14.0
Pillow==7.2.0
proto-plus==1.10.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytesseract==0.3.6
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
tesseract==0.1.3
tqdm==4.50.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
whitenoise==5.2.0
